app/search/pyproject.toml
[project]
name = "search"
version = "0.1.0"
description = ""
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "voyageai",
    "pymongo",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/search"]
